moved `find_all_of()` into `cppcheck` namespace and into `helpers.h`
