+++ test/helpers.h
@@ -92,6 +92,19 @@ public:
     static std::string getcode(Preprocessor &preprocessor, const std::string &filedata, const std::string &cfg, const std::string &filename, Suppressions *inlineSuppression = nullptr);
 };
 
+namespace cppcheck {
+    template<typename T>
+    std::size_t find_all_of(const std::string& str, T sub) {
+        std::size_t n = 0;
+        std::string::size_type pos = 0;
+        while ((pos = str.find(sub, pos)) != std::string::npos) {
+            ++pos;
+            ++n;
+        }
+        return n;
+    }
+}
+
 /* designated initialization helper
     Usage:
     struct S
